requirements.txt
Django==5.2.10
djangorestframework==3.15.2
mysqlclient==2.2.4
